requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
prefect>=3.0
google-api-python-client>=2.154.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
